requirements.txt
# Core dependencies for trading bot (exact versions for reproducibility)
alpaca-py==0.42.0
matplotlib==3.10.3
numpy==1.26.4
pandas==2.3.0
python-dateutil==2.9.0.post0
seaborn==0.13.2
tqdm==4.67.1
setuptools==80.9.0
python-telegram-bot==20.7
requests==2.31.0
PyYAML==6.0.1
pytest
boto3==1.34.131